requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=0.6,!=0.7,<1.0
Babel>=1.3
eventlet>=0.15.1
greenlet>=0.3.2
lockfile>=0.8
mysql-python
oslo.config>=1.2.0,<1.5
oslo.messaging>=1.3.0,<=1.4.1
pecan>=0.4.5
posix_ipc
python-keystoneclient>=0.4.2,<0.12
python-neutronclient>=2.3
python-novaclient==2.20
six>=1.6.0
SQLAlchemy>=0.7.8,!=0.9.5,<=0.9.99
sqlalchemy-migrate>=0.8.2,!=0.8.4,<=0.9.1
webob>=1.2.3
WSME>=0.6
